+++ CHANGES.rst
@@ -18,6 +18,7 @@ New features and enhancements
 Bug fixes
 ^^^^^^^^^
 * Fixed a bug with ``n_escore=-1`` in ``xclim.sdba.adjustment.NpdfTransform`` (:issue:`1515`, :pull:`1515`).
+* If chunked inputs are passed to indicators ``mean_radiant_temperature``  and ``potential_evapotranspiration``, subcalculations of the solar angle will also use the same chunks, instead of a single one of the same size as the data (:issue:`1536`, :pull:`1542`).
 
 Internal changes
 ^^^^^^^^^^^^^^^^